fixed home page styles and modified page nav
