+++ package.json
@@ -16,6 +16,7 @@
     "@popperjs/core": "^2.11.8",
     "gh-pages": "^6.3.0",
     "lucide-react": "^0.544.0",
+    "next": "^15.5.3",
     "react": "^19.1.1",
     "react-dom": "^19.1.1",
     "react-icons": "^5.5.0"
