+++ src/ProgressPage.jsx
@@ -1,10 +1,8 @@
-import SideMenu from "./components/side_menu/SideMenu.jsx"
 
 export default function ProgressPage() {
     return(
         <>
             <div className="main-content">
-                <SideMenu/>
                 <p>This page is in progress</p>
             </div>
         </>
