--- src/components/header/Header.jsx
@@ -1,22 +1,0 @@
-import logo_blue_short from "../../assets/blue_short.svg"
-
-function Header() {
-    return(
-        <div className="header-container">
-            <h2>getflicks</h2>
-            <p>&nbsp;powered by&nbsp;</p>
-            <div className="logo-container">
-                <a href="https://www.themoviedb.org/" target="blank">
-                <img id="logo" src={logo_blue_short} alt="TMDB logo blue_long_1"></img>
-                </a>
-            </div>
-            <div className="logo-container">
-                <a href="https://restcountries.com/#rest-countries" target="blank">
-                <h3>REST Countries</h3>
-                </a>
-            </div>
-        </div>
-    )
-}
-
-export default Header
