+++ src/components/header/navigator.css
@@ -1,0 +1,44 @@
+.page-nav {
+    height: 100%;
+    margin: 0px;
+}
+
+.page-nav .side-menu {
+    width: 100%;
+    border-left: 2px solid #0d253f;
+    border-right: 2px solid #0d253f;
+}
+
+.navigator {
+    display: flex;
+    height: 100%;
+    gap: 10px;
+}
+
+.button-side-menu, .button-side-menu-active {
+    width: 100%;
+    height: 100%;
+    border: none;
+    background-color: #0d253f;
+    color: white;
+    cursor: pointer;
+    font-size: large;
+}
+
+.button-side-menu-active {
+    background-color: #01b4e4;
+}
+
+.button-side-menu:hover {
+    background-color: hsl(211, 66%, 25%);
+}
+.button-side-menu:active {
+    background-color: hsl(212, 66%, 35%);
+}
+
+.button-side-menu-active:hover {
+    background-color: hsl(211, 66%, 25%);
+}
+.button-side-menu-active:active {
+    background-color: hsl(212, 66%, 35%);
+}
