+++ src/components/pages/body.css
@@ -1,6 +1,12 @@
 .main-content {
   width: 100%;
   display: grid;
+  grid-template-columns: 1fr;
+}
+
+.discover-main-content {
+  width: 100%;
+  display: grid;
   grid-template-columns: 1fr 10fr;
 }
