+++ src/components/pages/discover/DiscoverPage.jsx
@@ -3,7 +3,7 @@ import SideMenuDiscover from "../../side_menu/SideMenuDiscover";
 import MovieList from "./DiscoverMovieList";
 import { filterMovies } from "../../../../api/filterMovies";
 import { ActiveDisplayContext } from "../../../App";
-import { discoverLabel } from "../../side_menu/Navigator";
+import { discoverLabel } from "../../header/Navigator";
 
 export const FilteredMoviesContext = createContext()
 export const ActivePageContext = createContext()
@@ -40,7 +40,7 @@ function DiscoverPage() {
         <ParametersContext.Provider value={{setParameters}}>
         <ActivePageContext.Provider value={{activePage, toggleActivePage}}>
         <FilteredMoviesContext.Provider value={{filteredMovies, sendResults}}>
-            <div className="main-content">
+            <div className="discover-main-content">
                 <SideMenuDiscover/>
                 <MovieList/>
             </div>
