+++ src/components/pages/home/HomePage.tsx
@@ -1,11 +1,8 @@
-import SideMenu from "../../side_menu/SideMenu"
 import MovieSlider from "../MovieSlider"
-
-
-
-import { homeLabel } from "../../side_menu/Navigator"
+import { homeLabel } from "../../header/Navigator" 
 import { useEffect, useContext } from "react"
 import { ActiveDisplayContext } from "../../../App"
+import "./home_page.css"
 
 export const labelTopRated = "Top Rated"
 export const labelLatest = "Latest"
@@ -19,15 +16,12 @@ export default function HomePage(label:string) {
     },[])
 
     return(
-        <>
             <div className="main-content">
-                <SideMenu/>
                 <div className="home-page">
                     <MovieSlider label={labelTopRated}/>
                     <MovieSlider label={labelLatest}/>
                     <MovieSlider label={labelPopular}/>
                 </div>
             </div>
-        </>
     )
 }
