--- src/components/pages/home/SeeMoreButton.tsx
@@ -1,7 +1,0 @@
-export default function SeeMoreButton() {
-    return(
-        <>
-        <button>See More</button>
-        </>
-    )
-}
