+++ src/components/pages/movie_card.css
@@ -3,12 +3,12 @@
     min-width: 200px;
     max-width: 250px;
     height: fit-content;
-    min-height: 100%;
-    max-height: 500px;
+    max-height: 400px;
     display: flex;
     flex-direction: column;
     align-items: center;
     border-radius: 15px;
+    margin-bottom: 5px;
 }
 
 .movie-card:hover {
@@ -53,7 +53,7 @@
 .image {
     position: relative;
     width: 80%;
-    height: fit-content;
+    height: 100%;
 }
 
 .image .icon {
