+++ src/components/pages/movie_slider.css
@@ -1,0 +1,40 @@
+.slider-movie-list {
+    display: flex;
+    gap: 1rem;
+    list-style: none;
+    margin: 0;
+    padding: 0;
+    max-width: 1800px;
+    width: 100%;
+    overflow: scroll;
+}
+
+.slider-container {
+    width: 85vw;
+}
+
+.slider-header {
+    display: flex;
+    gap: 10px;
+    align-items: center;
+}
+
+.slider-header button {
+    height: 30px;
+    background-color: #0d253f;
+    padding: 10px;
+    color: white;
+    border: none;
+    border-radius: 15px;
+    display: flex;
+    align-items: center;
+    cursor: pointer;
+}
+
+.slider-header button:hover {
+    background-color: hsl(211, 66%, 30%);
+}
+
+.slider-header button:active {
+    background-color: hsl(211, 66%, 45%);
+}
