--- src/components/side_menu/SideMenu.jsx
@@ -1,13 +1,0 @@
-import Navigator from "./Navigator"
-import "./side_menu.css"
-
-
-function SideMenu() {
-    return(
-        <div className="side-menu">
-            <Navigator/>
-        </div>
-    )
-}
-
-export default SideMenu
