+++ src/components/side_menu/SideMenuDiscover.jsx
@@ -1,13 +1,10 @@
 import FilterSelector from "../pages/discover/FilterSelector"
-import ButtonSideMenu from "./Button_sm"
-import Navigator from "./Navigator"
 import "./side_menu.css"
 
 
 function SideMenuDiscover() {
     return(
         <div className="side-menu">
-            <Navigator/>
             <div className="discover-filters">
                 <FilterSelector/>
             </div>
