+++ tsconfig.json
@@ -9,7 +9,7 @@
     // See also https://aka.ms/tsconfig/module
     "module": "esnext",
     "target": "esnext",
-    "types": [],
+    "types": ["./declarations.d.ts"],
     // For nodejs:
     // "lib": ["esnext"],
     // "types": ["node"],
